+++ resources/lib/radiorai.py
@@ -2,7 +2,9 @@
 import urllib2
 import json
 
-class RadioRai:
+from .browser import Browser
+
+class RadioRai(Browser):
     def getChannels(self):
         url = "http://rai.it/dl/portaleRadio/popup/ContentSet-003728e4-db46-4df8-83ff-606426c0b3f5-json.html"
         response = json.load(urllib2.urlopen(url))
